Add MaterialDetails component
