+++ src/Style/index.scss
@@ -8,5 +8,5 @@
 @import '../containers/ProjectDetails/ProjectDetails';
 @import '../components/RoomsContainer/RoomsContainer';
 @import '../components/Room/Room';
-@import '../containers/RoomItems/RoomItems';
+@import '../containers/Materials/Materials';
 @import '../containers/NewRoom/NewRoom';
