+++ src/containers/Material/Material.js
@@ -1,64 +1,22 @@
-import React from 'react';
-import { Collapse } from 'antd';
-import NewMaterial from '../NewMaterial/NewMaterial';
+import React, { Component } from 'react';
 import EditMaterial from '../EditMaterial/EditMaterial';
 import DeleteConfirm from '../../components/DeleteConfirm/DeleteConfirm';
+import MaterialDetails from '../../components/MaterialDetails/MaterialDetails';
 
-const Panel = Collapse.Panel;
-
-function callback(key) {
-  console.log(key);
-}
-
-class Material extends React.Component {
-  state = { expanded: false };
-
-  renderMaterialTypes = () => {
-    let allCategories = []
-    this.props.materials.forEach(material => {
-      if (!allCategories.includes(material.element_type)) {
-        allCategories.push(material.element_type)
-      }
-    })
-
-    return allCategories.map((category, i) => {
-      return(
-        <Panel header={category} key={i}>
-          {this.renderMaterials(category)}
-        </Panel>
-      )
-    })
-  }
-
-  renderMaterials = (type) => {
-    let materials = this.props.materials.filter(material => (material.element_type === type))
-    return materials.map((material, i) => {
-      const { brand, manual_url, model_number, name, notes, quantity, unit_price, vendor } = material.material
-        return(
-          <div key={i} className='material'>
-            <p>{name}, {brand}</p>
-            <div className='material-control'>
-              <EditMaterial />
-              <DeleteConfirm type="material"/>
-            </div>
-          </div>
-        )
-    })
-  }
+class Material extends Component {
 
   render() {
-    return (
-      <div>
-        <div>
-          <Collapse onChange={callback}>
-            {this.renderMaterialTypes()}
-          </Collapse>
-        </div>
-        <div className='add-material-btn'>
-          <NewMaterial />
+    const { brand, name } = this.props.material
+    return(
+      <div key={name} className='material'>
+        <p>{name}, {brand}</p>
+        <div className='material-control'>
+          <MaterialDetails {...this.props.material}/>
+          <EditMaterial />
+          <DeleteConfirm type="material"/>
         </div>
       </div>
-    );
+    )
   }
 }
