+++ src/containers/Materials/Materials.test.js
@@ -1,0 +1,73 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import Materials from './Materials';
+
+describe('RoomItems', () => {
+  let wrapper;
+  let mockMaterials;
+  
+  beforeEach(() => {
+    mockMaterials = [
+      {
+          "id": 1,
+          "element_type": "Flooring",
+          "material": {
+              "id": 1,
+              "name": "Material 1",
+              "model_number": "abc1",
+              "brand": "Kenmoore",
+              "vendor": "HD",
+              "manual_url": null,
+              "notes": null,
+              "quantity": null,
+              "unit_price": null
+          }
+      },
+      {
+          "id": 2,
+          "element_type": "Flooring",
+          "material": {
+              "id": 2,
+              "name": "Material 2",
+              "model_number": "abc2",
+              "brand": "Kenmoore",
+              "vendor": "HD",
+              "manual_url": null,
+              "notes": null,
+              "quantity": null,
+              "unit_price": null
+          }
+      },
+      {
+        "id": 3,
+        "element_type": "Fixtures",
+        "material": {
+            "id": 3,
+            "name": "Material A",
+            "model_number": "abc1",
+            "brand": "GE",
+            "vendor": "HD",
+            "manual_url": null,
+            "notes": null,
+            "quantity": null,
+            "unit_price": null
+        }
+    }
+  ]
+    wrapper = shallow(<Materials materials={mockMaterials} />)
+  });
+
+  it('should match snapshot', () => {
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should collect all categories when renderMaterialTypes is called', () => {
+    let result = wrapper.instance().renderMaterialTypes()
+    expect(result.length).toBe(2)
+  });
+
+  it('should render materials when renderMaterials is called', () => {
+    let result = wrapper.instance().renderMaterials('Flooring')
+    expect(result.length).toBe(2)
+  });
+});
